Object touch interaction
Added functionality for exit object
